refactor(#3104): extract source-scan predicates from codegen/index.ts
Slice 1 of the src/codegen/index.ts (#3104) god-file decomposition. Lift
the 5 whole-program AST pre-scan predicates (sourceContainsClass,
sourceContainsDelete, sourceHasDynamicTaConstruct,
sourceContainsBindingPattern, sourceOverridesArrayIterator) verbatim out
of index.ts (block 426-659) into a new sibling module
src/codegen/source-scan-predicates.ts.

index.ts shrinks -235 lines. Bodies moved verbatim; predicates have no
dependency on CodegenContext (only ts/forEachChild + TYPED_ARRAY_NAMES).
Imported back into index.ts; sourceOverridesArrayIterator also re-exported
for the external test consumer. The index.ts <-> predicates import edge is
load-safe (predicates run only at codegen time, never at module init).

REFACTOR, zero behavior change. tsc --noEmit clean; check:loc-budget green;
targeted vitest (issue-1719-s1, issue-1364b-class-method-delete): 21/21 pass.

Original extraction by dev-l; recovered, validated, and landed by dev-k
after a box-recovery from swap-death.

✓
